+++ src/jsontoolkit/include/jsontoolkit/json_number.h
@@ -32,9 +32,10 @@ const char token_number_seven = '\u0037';
 const char token_number_eight = '\u0038';
 const char token_number_nine = '\u0039';
 
-auto parse(const std::string_view &input) -> std::variant<std::int64_t, double>;
+auto parse(const std::string_view &input)
+    -> std::variant<std::int64_t, long double>;
 auto stringify(std::int64_t value) -> std::string;
-auto stringify(double value) -> std::string;
+auto stringify(long double value) -> std::string;
 
 } // namespace sourcemeta::jsontoolkit::Number
