+++ test/jsontoolkit/number_test.cc
@@ -12,7 +12,7 @@ TEST(Number, set_negative_integer) {
 TEST(Number, set_negative_real) {
   sourcemeta::jsontoolkit::JSON document{"true"};
   EXPECT_TRUE(document.is_boolean());
-  document = -4.3;
+  document = static_cast<long double>(-4.3);
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
 }
@@ -20,7 +20,7 @@ TEST(Number, set_negative_real) {
 TEST(Number, set_negative_integral_real) {
   sourcemeta::jsontoolkit::JSON document{"true"};
   EXPECT_TRUE(document.is_boolean());
-  document = -4.0;
+  document = static_cast<long double>(-4.0);
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
 }
@@ -36,7 +36,7 @@ TEST(Number, set_positive_integer) {
 TEST(Number, set_positive_real) {
   sourcemeta::jsontoolkit::JSON document{"true"};
   EXPECT_TRUE(document.is_boolean());
-  document = 4.3;
+  document = static_cast<long double>(4.3);
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
 }
@@ -44,7 +44,7 @@ TEST(Number, set_positive_real) {
 TEST(Number, set_positive_integral_real) {
   sourcemeta::jsontoolkit::JSON document{"true"};
   EXPECT_TRUE(document.is_boolean());
-  document = 4.0;
+  document = static_cast<long double>(4.0);
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
 }
@@ -55,7 +55,7 @@ TEST(Number, string_integer_zero) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 0);
   EXPECT_EQ(document, static_cast<std::int64_t>(0));
-  EXPECT_EQ(document, static_cast<double>(0));
+  EXPECT_EQ(document, static_cast<long double>(0));
 }
 
 TEST(Number, string_integer_positive_single_digit_integer) {
@@ -64,7 +64,7 @@ TEST(Number, string_integer_positive_single_digit_integer) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 3);
   EXPECT_EQ(document, static_cast<std::int64_t>(3));
-  EXPECT_EQ(document, static_cast<double>(3));
+  EXPECT_EQ(document, static_cast<long double>(3));
 }
 
 TEST(Number, string_integer_negative_single_digit_integer) {
@@ -73,7 +73,7 @@ TEST(Number, string_integer_negative_single_digit_integer) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), -3);
   EXPECT_EQ(document, static_cast<std::int64_t>(-3));
-  EXPECT_EQ(document, static_cast<double>(-3));
+  EXPECT_EQ(document, static_cast<long double>(-3));
 }
 
 TEST(Number, string_integer_positive_long_integer) {
@@ -82,7 +82,7 @@ TEST(Number, string_integer_positive_long_integer) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 12345);
   EXPECT_EQ(document, static_cast<std::int64_t>(12345));
-  EXPECT_EQ(document, static_cast<double>(12345));
+  EXPECT_EQ(document, static_cast<long double>(12345));
 }
 
 TEST(Number, string_integer_negative_long_integer) {
@@ -91,7 +91,7 @@ TEST(Number, string_integer_negative_long_integer) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), -12345);
   EXPECT_EQ(document, static_cast<std::int64_t>(-12345));
-  EXPECT_EQ(document, static_cast<double>(-12345));
+  EXPECT_EQ(document, static_cast<long double>(-12345));
 }
 
 TEST(Number, string_integer_minus_zero) {
@@ -100,7 +100,7 @@ TEST(Number, string_integer_minus_zero) {
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 0);
   EXPECT_EQ(document, static_cast<std::int64_t>(0));
-  EXPECT_EQ(document, static_cast<double>(0));
+  EXPECT_EQ(document, static_cast<long double>(0));
 }
 
 TEST(Number, invalid_string_integer_trailing_minus) {
@@ -178,7 +178,7 @@ TEST(Number, trailing_zero_positive) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.5);
-  EXPECT_EQ(document, 1.5);
+  EXPECT_EQ(document, static_cast<long double>(1.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1));
   EXPECT_FALSE(document == static_cast<std::int64_t>(2));
 }
@@ -188,7 +188,7 @@ TEST(Number, trailing_zero_negative) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.5);
-  EXPECT_EQ(document, -1.5);
+  EXPECT_EQ(document, static_cast<long double>(-1.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-2));
 }
@@ -198,7 +198,7 @@ TEST(Number, single_left_digit_positive_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.5);
-  EXPECT_EQ(document, 1.5);
+  EXPECT_EQ(document, static_cast<long double>(1.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1));
   EXPECT_FALSE(document == static_cast<std::int64_t>(2));
 }
@@ -208,7 +208,7 @@ TEST(Number, single_left_digit_negative_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.5);
-  EXPECT_EQ(document, -1.5);
+  EXPECT_EQ(document, static_cast<long double>(-1.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-2));
 }
@@ -218,7 +218,7 @@ TEST(Number, leading_decimal_zero) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0005);
-  EXPECT_EQ(document, 1.0005);
+  EXPECT_EQ(document, static_cast<long double>(1.0005));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1));
   EXPECT_FALSE(document == static_cast<std::int64_t>(2));
 }
@@ -228,7 +228,7 @@ TEST(Number, multi_left_digit_positive_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1234.5);
-  EXPECT_EQ(document, 1234.5);
+  EXPECT_EQ(document, static_cast<long double>(1234.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1234));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1235));
 }
@@ -238,7 +238,7 @@ TEST(Number, multi_left_digit_negative_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1234.5);
-  EXPECT_EQ(document, -1234.5);
+  EXPECT_EQ(document, static_cast<long double>(-1234.5));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1235));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1234));
 }
@@ -248,7 +248,7 @@ TEST(Number, long_positive_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1234.56789);
-  EXPECT_EQ(document, 1234.56789);
+  EXPECT_EQ(document, static_cast<long double>(1234.56789));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1234));
   EXPECT_FALSE(document == static_cast<std::int64_t>(1235));
 }
@@ -258,7 +258,7 @@ TEST(Number, long_negative_real) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1234.56789);
-  EXPECT_EQ(document, -1234.56789);
+  EXPECT_EQ(document, static_cast<long double>(-1234.56789));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1235));
   EXPECT_FALSE(document == static_cast<std::int64_t>(-1234));
 }
@@ -278,7 +278,7 @@ TEST(Number, single_digit_positive_real_integer) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0);
-  EXPECT_EQ(document, 1.0);
+  EXPECT_EQ(document, static_cast<long double>(1.0));
   EXPECT_EQ(document, static_cast<std::int64_t>(1));
 }
 
@@ -287,7 +287,7 @@ TEST(Number, single_digit_positive_real_integer_trailing_zero) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0);
-  EXPECT_EQ(document, 1.0);
+  EXPECT_EQ(document, static_cast<long double>(1.0));
   EXPECT_EQ(document, static_cast<std::int64_t>(1));
 }
 
@@ -296,7 +296,7 @@ TEST(Number, single_digit_negative_real_integer) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.0);
-  EXPECT_EQ(document, -1.0);
+  EXPECT_EQ(document, static_cast<long double>(-1.0));
   EXPECT_EQ(document, static_cast<std::int64_t>(-1));
 }
 
@@ -466,7 +466,7 @@ TEST(Number, exponential_notation_plus_after_e) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
   EXPECT_EQ(document, static_cast<std::int64_t>(300));
-  EXPECT_EQ(document, static_cast<double>(300));
+  EXPECT_EQ(document, static_cast<long double>(300));
 }
 
 // From https://en.wikipedia.org/wiki/Scientific_notation
@@ -477,7 +477,7 @@ TEST(Number, exponential_notation_integer_1_upper) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 2.0);
   EXPECT_EQ(document, static_cast<std::int64_t>(2));
-  EXPECT_EQ(document, static_cast<double>(2));
+  EXPECT_EQ(document, static_cast<long double>(2));
 }
 
 TEST(Number, exponential_notation_integer_2_upper) {
@@ -486,7 +486,7 @@ TEST(Number, exponential_notation_integer_2_upper) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
   EXPECT_EQ(document, static_cast<std::int64_t>(300));
-  EXPECT_EQ(document, static_cast<double>(300));
+  EXPECT_EQ(document, static_cast<long double>(300));
 }
 
 TEST(Number, exponential_notation_integer_3_upper) {
